Fix save actions (and rename `remove_unicode_ligatures` to `replace_unicode_ligatures`)
